+++ api/lib/application/sco-organization-learners/sco-organization-learner-controller.js
@@ -26,23 +26,6 @@ const generateUsername = async function (request, h, dependencies = { scoOrganiz
     .code(200);
 };
 
-const updatePassword = async function (request, h, dependencies = { scoOrganizationLearnerSerializer }) {
-  const payload = request.payload.data.attributes;
-  const userId = request.auth.credentials.userId;
-  const organizationId = payload['organization-id'];
-  const organizationLearnerId = payload['organization-learner-id'];
-
-  const scoOrganizationLearner = await usecases.updateOrganizationLearnerDependentUserPassword({
-    userId,
-    organizationId,
-    organizationLearnerId,
-  });
-
-  return h
-    .response(dependencies.scoOrganizationLearnerSerializer.serializeCredentialsForDependent(scoOrganizationLearner))
-    .code(200);
-};
-
 const generateUsernameWithTemporaryPassword = async function (
   request,
   h,
@@ -88,7 +71,6 @@ const batchGenerateOrganizationLearnersUsernameWithTemporaryPassword = async fun
 
 const libScoOrganizationLearnerController = {
   generateUsername,
-  updatePassword,
   generateUsernameWithTemporaryPassword,
   batchGenerateOrganizationLearnersUsernameWithTemporaryPassword,
 };
