+++ api/tests/unit/domain/usecases/update-organization-learner-dependent-user-password_test.js
@@ -1,4 +1,4 @@
-import { updateOrganizationLearnerDependentUserPassword } from '../../../../lib/domain/usecases/update-organization-learner-dependent-user-password.js';
+import { updateOrganizationLearnerDependentUserPassword } from '../../../../src/prescription/organization-learner/domain/usecases/update-organization-learner-dependent-user-password.js';
 import { UserNotFoundError } from '../../../../src/shared/domain/errors.js';
 import { UserNotAuthorizedToUpdatePasswordError } from '../../../../src/shared/domain/errors.js';
 import { catchErr, expect, sinon } from '../../../test-helper.js';
